services/scrappers/app.py
from fastapi import FastAPI, Query
from pydantic import BaseModel
import requests
from bs4 import BeautifulSoup
from dateutil import parser as dateparser

app = FastAPI(title="Scrapers")

class Result(BaseModel):
    source: str
    url: str
    title: str
    snippet: str
    published_at: str

# Simple RSS or page scraper examples
def scrape_bmc():
    # placeholder: BMC may have RSS; we'll simulate
    return [
        {"source": "BMC", "url":"https://bmc.gov.in/sample1", "title":"BMC Alert: Storm", "snippet":"BMC issues alert for ...", "published_at":"2025-11-01T10:00:00Z"}
    ]

def scrape_mumbai_police():
    return [
        {"source":"Mumbai Police","url":"https://mumbaipolice.gov.in/sample1", "title":"Mumbai Police advisory", "snippet":"Traffic diversions ...", "published_at":"2025-11-01T09:00:00Z"}
    ]

@app.get("/search")
async def search(q: str = Query(..., description="query")):
    # In production: call site-specific scrapers and fulltext search
    results = []
    results.extend(scrape_bmc())
    results.extend(scrape_mumbai_police())
    # naive filter
    filtered = [r for r in results if q.lower().split()[0] in (r['title'] + r['snippet']).lower()]
    return {"results": filtered or results}
