services/scrappers/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
python-dateutil
